refactor: extract replay CLI commands into internal/replaytool
Move the offline replay/replay-range developer commands out of
internal/cli into a dedicated internal/replaytool package. The replay
code (~2.8k LOC) had no coupling to the rest of cli beyond root command
registration; the new package exposes NewCommands(), which cli/root.go
wires onto the root command. Pure file moves plus a small registration
shim and a doc update.
