+++ internal/replaytool/replay_checkpoint.go
@@ -1,4 +1,4 @@
-package cli
+package replaytool
 
 import (
 	"bufio"
