+++ internal/replaytool/replay_findings.go
@@ -1,4 +1,4 @@
-package cli
+package replaytool
 
 import (
 	"encoding/hex"
