+++ internal/replaytool/replay_findings_test.go
@@ -1,4 +1,4 @@
-package cli
+package replaytool
 
 import (
 	"bufio"
