+++ internal/replaytool/replay_range_helpers_test.go
@@ -1,4 +1,4 @@
-package cli
+package replaytool
 
 import (
 	"encoding/hex"
