+++ internal/replaytool/replay_range_test.go
@@ -1,4 +1,4 @@
-package cli
+package replaytool
 
 import (
 	"os"
