+++ internal/replaytool/replay_reconstruct_test.go
@@ -1,4 +1,4 @@
-package cli
+package replaytool
 
 import (
 	"encoding/hex"
