+++ internal/replaytool/replay_state_source.go
@@ -1,4 +1,4 @@
-package cli
+package replaytool
 
 import (
 	"context"
