+++ internal/replaytool/replay_state_source_test.go
@@ -1,4 +1,4 @@
-package cli
+package replaytool
 
 import (
 	"context"
